Refactoring - mocing ChooseStripe to services.stripe package.
