+++ membership-attribute-service/app/components/TouchpointComponents.scala
@@ -22,7 +22,7 @@ import org.http4s.Uri
 import scalaz.std.scalaFuture._
 import services._
 import services.salesforce.{ContactRepository, ContactRepositoryWithMetrics, CreateScalaforce, SimpleContactRepository}
-import services.stripe.{BasicStripeService, BasicStripeServiceWithMetrics, HttpBasicStripeService}
+import services.stripe.{BasicStripeService, BasicStripeServiceWithMetrics, ChooseStripe, HttpBasicStripeService}
 import services.subscription.{SubscriptionService, SubscriptionServiceWithMetrics, ZuoraSubscriptionService}
 import services.zuora.rest.{SimpleClient, SimpleClientZuoraRestService, ZuoraRestService, ZuoraRestServiceWithMetrics}
 import services.zuora.soap.{SimpleZuoraSoapService, ZuoraSoapService, ZuoraSoapServiceWithMetrics}
